Added radio buttons to rubric creation for half/quarter point grade increments, added dynamic coloring for rubric levels, and changed some code to handle doubles instead of ints for scores.

+++ OSBLEPlus/trunk/development/OSBLE.Library/Models/AbstractCourses/Course/Rubrics/RubricEvaluation.cs
@@ -106,7 +106,7 @@ namespace OSBLE.Models.Courses.Rubrics
                         double currentWeight = critEval.Criterion.Weight;
                         sumOfWeights += currentWeight;
 
-                        int? currentScore = critEval.Score;
+                        double? currentScore = critEval.Score;
 
                         if (currentScore.HasValue)
                         {
