+++ OSBLEPlus/trunk/development/OSBLEPlus.Database/Schema/Tables/CriterionEvaluations.sql
@@ -2,7 +2,7 @@
     [CriterionID]        INT             NOT NULL,
     [ID]                 INT             IDENTITY (1, 1) NOT NULL,
     [RubricEvaluationID] INT             NOT NULL,
-    [Score]              INT             NULL,
+    [Score]              FLOAT             NULL,
     [Comment]            NVARCHAR (4000) NULL,
     CONSTRAINT [PK_dbo.CriterionEvaluations] PRIMARY KEY CLUSTERED ([ID] ASC),
     CONSTRAINT [FK_dbo.CriterionEvaluations_dbo.Criteria_CriterionID] FOREIGN KEY ([CriterionID]) REFERENCES [dbo].[Criteria] ([ID]) ON DELETE CASCADE,
